refactor: 객체 themes 사용
