+++ src/components/Header/index.tsx
@@ -1,6 +1,6 @@
 import { Container } from '@/components/common/layouts/Container';
-import { GetTheme } from '@/components/Header/GetTheme';
 import Nav from '@/components/Header/Nav';
+import { themes } from '@/components/Header/themes';
 import Title from '@/components/Header/Title';
 
 interface HeaderProps {
@@ -8,7 +8,7 @@ interface HeaderProps {
 }
 
 const Header = ({ themeKey }: HeaderProps) => {
-  const theme = themeKey ? GetTheme()[themeKey] : null;
+  const theme = themeKey ? themes[themeKey] : null;
   const backgroundColor = theme ? theme.backgroundColor : 'transparent';
 
   return (
