+++ src/pages/MainPage/Theme.tsx
@@ -2,11 +2,10 @@ import { useNavigate } from 'react-router-dom';
 
 import { Image } from '@/components/common/Image';
 import { Grid } from '@/components/common/layouts/Grid';
-import { GetTheme } from '@/components/Header/GetTheme';
+import { themes } from '@/components/Header/themes';
 
 const ThemeComponent = () => {
   const navigate = useNavigate();
-  const themes = GetTheme();
 
   const handleClick = (themeKey: string) => {
     navigate(`/theme/${themeKey}`);
